refactor: remove builder and replace with usecase
