+++ tests/simulate/test_view.py
@@ -1,11 +1,11 @@
 from typing import TypedDict
 
 from simaple.core.base import ActionStat
-from simaple.simulate.builder import EngineBuilder
 from simaple.simulate.component.base import Component, view_method
 from simaple.simulate.core.base import Entity
 from simaple.simulate.core.store import AddressedStore, ConcreteStore
 from simaple.simulate.global_property import GlobalProperty
+from simaple.simulate.usecase import Usecase
 
 
 class SomeEntity(Entity):
@@ -32,10 +32,9 @@ def test_view():
     global_property = GlobalProperty(ActionStat())
     global_property.install_global_properties(store)
 
-    runtime = (
-        EngineBuilder(store=store)
-        .add_component(ViewTestComponent(id="dummy", name="test_component"))
-        .build_simulation_runtime()
-    )
+    usecase = Usecase()
+    usecase.use_component(ViewTestComponent(id="dummy", name="test_component"))
+
+    runtime = usecase.create_simulation_runtime(store=store)
 
     assert runtime.get_viewer()("test_component.naming") == str(3 + 5)
